refactor: make album/artist/track fixtures available everywhere.
